pyproject.toml
[project]
name = "franka_assembly"
version = "0.0.1"
authors = [
  { name="Wenjun Liu", email="wenjunliu@ust.hk" },
]
dependencies = [
#    "pytest >= 8.3.0",
    "trimesh >= 4.6.4",
#    "python-fcl >= 0.7.0.8",
    "scipy >= 1.15.2",
#    "warp-lang >= 1.6.2",
#    "usd-core >= 25.2",
#    "polyscope >= 2.3.0",
    "hatchling >= 1.27.0",
#    "pytorch-kinematics >= 0.7.5",
    "compas_eve >= 0.5.0",
    "franky-control>= 1.0.2"

]

description = "A package for conducting research using robots for assembly"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.urls]
Homepage = "https://github.com/wenjun-dbt/FrankaAssembly"
Issues = "https://github.com/wenjun-dbt/FrankaAssembly/issues"